+++ finite-automata-designer/src/app/components/editor/export/ExportContainer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useRef, useState } from "react";
-import { toggle_visiblity } from "../../../../public/scripts/canvasUtil/canvasUtil";
+import { toggle_visiblity } from "../../../../../public/scripts/canvasUtil/canvasUtil";
 
 export default function ExportContainer(){
